// apps/worker/ui/src/api/types.ts
export type ExampleRecord = {
  id?: number;
  created_at?: string | null;
  updated_at?: string | null;
  jp_text: string;
  kr_text: string;
  context?: string | null;
  prompt?: string | null;
  negative_prompt?: string | null;
  context_embedding?: number[] | null;
  text_embedding?: number[] | null;
  prompt_embedding?: number[] | null;
  jp_words?: number[] | null;
  audios?: number[] | null;
  error_reports?: number[] | null;
};
// ...
export type ImageRecord = {
  id?: number;
  created_at?: string | null;
  updated_at?: string | null;
  prompt?: string | null;
  negative_prompt?: string | null;
  prompt_embedding?: number[] | null;
  object_key?: string | null;
};